Add pyramid heightmap script and demo world
This adds a python script to the media/materials/textures
folder that can be used to generate a png file with a
square grayscale gradient that corresponds to a pyramid
heightmap.

Due to #2838, the triangulation pattern differs between
OGRE and ODE, so some contact behavior can look wrong
in the places where this visualization differs.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>